% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accidents2017}
\alias{accidents2017}
\title{Road Traffic Accident Analysis for Barcelona, 2017.}
\format{
A data frame with 10335 rows and 8 variables:
\describe{
\item{Id}{id that uniquely identifies the accident.}
\item{Street}{The street where the accident occurred.}
\item{datetime}{The timestamp when the accident occurred.}
\item{Mild.injuries}{The number of mild injuries caused by the accident.}
\item{Serious.injuries}{The number of serious injuries caused by the accident.}
\item{Victims}{The total number of victims involved in the accident.}
\item{Vehicles.involved}{The number of vehicles involved in the accident.}
\item{geometry}{The location where the accident happened.}
}
}
\source{
\url{https://www.kaggle.com/datasets/xvivancos/barcelona-data-sets}
}
\usage{
data(accidents2017)
}
\description{
A data set containing road traffic accidents.
}
\keyword{datasets}
