% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocationAnalysis.R
\name{Display_AnimatedHeatMap}
\alias{Display_AnimatedHeatMap}
\title{Animated Heatmap for the accidents data for 24 hours}
\usage{
Display_AnimatedHeatMap(data_sf)
}
\arguments{
\item{data}{A data frame containing the car accident data}
}
\value{
A map that shows the changes of the accident intensities for 24 hours
}
\description{
This function displays the animated heatmap for each hours in 24 hours.
}
