% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocationAnalysis.R
\name{Display_clusteraccident_serious}
\alias{Display_clusteraccident_serious}
\title{Display the cluster of the traffic car accident data}
\usage{
Display_clusteraccident_serious(accidents_data)
}
\arguments{
\item{accidents_data}{A data frame containing the car accident data}
}
\value{
a map displaying the clustering the serious accidents based on the area and the counts of the accidents
}
\description{
This function displays and calculates clustering the serious injuries and count the number of accidents in the cluster.
}
