% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocationAnalysis.R
\name{Display_Serious_injuries}
\alias{Display_Serious_injuries}
\title{Displays the serious injuries occurred in the area}
\usage{
Display_Serious_injuries(accidents_data)
}
\arguments{
\item{accidents_data}{A data frame containing the car accident data}
}
\value{
a map with serious accidents and the number of serious accidents in the area
}
\description{
This function Displays serious accidents for the car accident data.
}
