% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocationAnalysis.R
\name{Display_study_area}
\alias{Display_study_area}
\title{Displays the study area for this project}
\usage{
Display_study_area()
}
\value{
Leaflet map with zoom 12 of the study area,barcelona
}
\description{
This function displays the study area for this data analysis in this case Barcelona.Spain
}
