% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{perform_count_accidents}
\alias{perform_count_accidents}
\title{Perform Road Traffic Accidents Analysis}
\usage{
perform_count_accidents(accidents)
}
\arguments{
\item{accidents}{A data frame containing the road traffic accidents dataset.}
}
\value{
A list of analysis results.
}
\description{
This function performs various analyses on the road traffic accidents dataset.
}
