% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspotAnalysis.R
\name{perform_hotspot_count}
\alias{perform_hotspot_count}
\title{Perform hotspot analysis of counting the number of points in each cell grid}
\usage{
perform_hotspot_count(data)
}
\arguments{
\item{data}{A data frame containing the car accident data}
}
\value{
A list of hotspot counts
}
\description{
Count the number of points in each cell of a regular grid. Cell size can be set by the user or chosen automatically.
}
