% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspotAnalysis.R
\name{Perform_hotspot_gistar}
\alias{Perform_hotspot_gistar}
\title{Perform hotspot analysis for the car accident}
\usage{
Perform_hotspot_gistar(data)
}
\arguments{
\item{data}{A data frame containing the car accident data}
}
\value{
Accidents data with calculated density
}
\description{
This function calculates the Getis–Ord Gi* statistic for each cell in a regular grid, while optionally estimating kernel density..
}
