% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspotAnalysis.R
\name{perform_hotspot_kde}
\alias{perform_hotspot_kde}
\title{Perform kernel density estimation  hotpots analysis}
\usage{
perform_hotspot_kde(data)
}
\arguments{
\item{data}{A data frame containing the car accident data}
}
\value{
A named list of summary statistics
}
\description{
Estimate kernel density for each cell in a regular grid. Cell size and bandwidth can be set by the user or chosen automatically.
}
