% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{perform_monthly_analysis}
\alias{perform_monthly_analysis}
\title{Perform Monthly Analysis on Car Accident Data}
\usage{
perform_monthly_analysis(accidents_data)
}
\arguments{
\item{accidents_data}{A data frame containing car accident data with a 'datetime' column representing the date and time of each accident.}
}
\value{
A summarized data frame providing insights into the number of accidents that occurred in each month.
}
\description{
This function performs monthly analysis on car accident data, enabling you to explore accident patterns on a monthly basis.
}
