% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{perform_percentage_for_mid_serious}
\alias{perform_percentage_for_mid_serious}
\title{Perform accident type analysis for comparing the mid and serious injuries}
\usage{
perform_percentage_for_mid_serious(accidents_data)
}
\arguments{
\item{accidents_data}{A data frame containing the car accident data}
}
\value{
A pie chart visualizing the percentage of mild and serious injuries
}
\description{
This function performs accident type analysis on the car accident data and visualizes the percentage of mild and serious injuries using a pie chart.
}
