% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{perform_rateofinjury_bydayofweek}
\alias{perform_rateofinjury_bydayofweek}
\title{Perform the rate of the accident by the day of the week}
\usage{
perform_rateofinjury_bydayofweek(accidents_data)
}
\arguments{
\item{data}{A data frame containing the car accident data}
}
\value{
A data frame summarizing the accidents based on the chosen time dimension
}
\description{
This function performs comparison of the rate of accidents between serious and mid injuries
}
