% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{perform_vehicledInvolved}
\alias{perform_vehicledInvolved}
\title{Perform type of accident analysis with vichles involed in the accident}
\usage{
perform_vehicledInvolved(accidents_data)
}
\arguments{
\item{data}{A data frame containing the car accident data}
}
\value{
A data frame summarizing the accidents based on the chosen time dimension
}
\description{
This function performs time analysis on the car accident data, allowing you to analyze the accidents
based on different time dimensions.
}
