% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_analysis.R
\name{perform_weekdays_analysis}
\alias{perform_weekdays_analysis}
\title{Perform time analysis on car accident data in the weekdays}
\usage{
perform_weekdays_analysis(accidents_data)
}
\arguments{
\item{accidents_data}{A data frame containing the car accident data}
}
\value{
A data frame summarizing the accidents based on the chosen time dimension
}
\description{
This function performs time analysis on the car accident data, allowing you to analyze the accidents
based on different time dimensions.
}
