% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{perform_percentage_for_mid_serious}
\alias{perform_percentage_for_mid_serious}
\title{Perform accident type analysis and percenatge}
\usage{
perform_percentage_for_mid_serious(accidents_data)
}
\arguments{
\item{data}{A data frame containing the car accident data}
}
\value{
A data frame summarizing the accidents based on the chosen time dimension
}
\description{
This function performs accident type analysis on the car accident data
}
